feat: 새로운 애플리케이션 구조 및 초기화 로직 구현
- BootstrapApplication 클래스를 추가하여 애플리케이션 초기화 로직 통합
- DOMManager, EventManager, UIUpdater, CalculationEngine, PromotionManager, ShoppingCartState 클래스를 활용하여 상태 관리 및 UI 업데이트 로직 구성
- main.basic.js에서 기존 구조를 새로운 BootstrapApplication으로 변경하여 초기화 과정 간소화
- 전체적인 코드 리팩토링으로 유지보수성 및 확장성 향상

✅ 새로운 구조로 애플리케이션의 초기화 및 상태 관리 효율성 개선
